GEODE-3903: Moved serialization package to under protocol.serialization moved META-INF services from protobuf to client-protocol
