+++ geode-client-protocol/src/main/java/org/apache/geode/internal/protocol/serialization/registry/exception/CodecNotRegisteredForTypeException.java
@@ -12,25 +12,16 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.internal.serialization;
+package org.apache.geode.internal.protocol.serialization.registry.exception;
 
 import org.apache.geode.annotations.Experimental;
 
 /**
- * This interface converts a particular type to and from its binary representation.
- *
- * NOTE: it is expected that T will be one of the serialization types in @{@link SerializationType}.
- *
- * @param <T> the type this codec knows how to convert
+ * This indicates we're attempting to handle a type for which we don't have a registered codec.
  */
 @Experimental
-public interface TypeCodec<T> {
-  T decode(byte[] incoming);
-
-  byte[] encode(T incoming);
-
-  /**
-   * @return the SerializationType corresponding to T
-   */
-  SerializationType getSerializationType();
+public class CodecNotRegisteredForTypeException extends Exception {
+  public CodecNotRegisteredForTypeException(String message) {
+    super(message);
+  }
 }
