+++ geode-client-protocol/src/main/resources/META-INF/services/org.apache.geode.internal.protocol.serialization.TypeCodec
@@ -1,0 +1,1 @@
+org.apache.geode.internal.protocol.serialization.codec.JSONCodec
