--- geode-protobuf/src/main/resources/META-INF/services/org.apache.geode.internal.serialization.TypeCodec
@@ -1,1 +1,0 @@
-org.apache.geode.internal.serialization.codec.JSONCodec
