+++ geode-protobuf/src/test/java/org/apache/geode/internal/protocol/protobuf/operations/OperationHandlerJUnitTest.java
@@ -21,7 +21,7 @@ import org.junit.experimental.categories.Category;
 
 import org.apache.geode.cache.Cache;
 import org.apache.geode.internal.protocol.operations.OperationHandler;
-import org.apache.geode.internal.serialization.SerializationService;
+import org.apache.geode.internal.protocol.serialization.SerializationService;
 import org.apache.geode.test.junit.categories.UnitTest;
 
 @Category(UnitTest.class)
